chore: reconcile extractor rfc-validator.ts
